Fix channel disabling and endpoint recovery
The HAL_HCD_HC_Halt() function doesn't actually halt a channel correctly,
so this fix replaces those calls with code that does. This is necessary
both when freeing the bulk endpoints when a USB thumb drive is disconnected,
but also when performing error recovery on a channel that is blocked in a
NAK state. The HAL_HCD_HC_Init() function doesn't return the channel to an
active state itself, so there's also code added to do that. Finally, the
endpoint object must be returned to USB_TYPE_IDLE at the end of such an
error recovery.