Add guards against uninitialized host pointer
An older fix (commit 72d0aa6) introduced the possiblity of dereferencing an
uninitialized host pointer by moving the getHostInst() call from USBHostMSD()'s
constructor to connect(). This is now solved by adding guards against an
uninitialized host pointer in checkResult(), SCSITransfer() and getMaxLun().
The practical implication of the bug was that mounting before connecting led
to a crash, but now just leads to an error message from the mount call.