Fix I/O errors when using slow USB thumb drives
Increases the MAX_NOTREADY_RETRY from 5 to 50000. The previous value (5) was not
nearly enough to handle the slow writes to some USB thumb drives. While it worked
with some models, it failed all the time with others. 50000 might be a bit on the
high side, but it adds margin since we can't possibly test all the models out there.