Fix missing default Mbed configuration
Parts of the library don't know about the default Mbed configuration. If only
the redefines are removed, Callback.h is included already in USBHostConf.h,
but without any default Mbed configuration, so we get the trivial version of
call(). Next, Arduino.h is included, which includes the Mbed configuration
indirectly, but the damage has already been done. Over in USBHostSerial.h,
USBHostConf.h is included. After that, Callback.h is included in
USBHostSerial.h, and we have the correct Mbed configuration in place, but
since it's already been included once, the correct version of call() is never
actually used. So we get two versions of call() again, which is undefined
behavior, and a crash.